// src/types/Vendor.ts
export interface Vendor {
    description: string | null;
    email: string | null;
    capacity: number | null;
    id_image: number | null;
    id_person: number;
    id_status: number | null;
    id_vendor: number;
    id_video: number | null;
    link: string | null;
    name: string;
    requests_open: boolean | null;
    utc_timestamp: string;
  }
  